♻️ Refactor: #10 접기기능 모듈화 작업
