+++ src/components/modules/getVertexFromPosition.js
@@ -1,0 +1,13 @@
+import * as THREE from 'three';
+
+const getVertexFromPosition = (allPositions, i) => {
+  if (!allPositions.isBufferAttribute) {
+    return allPositions[i];
+  } else {
+    const vertex = new THREE.Vector3();
+    vertex.fromBufferAttribute(allPositions, i);
+    return vertex;
+  }
+};
+
+export { getVertexFromPosition };
